Better organize Sharepoint related methods
