+++ server/src/document/document.module.ts
@@ -3,11 +3,13 @@ import { ConfigModule } from '../config/config.module';
 import { DocumentController } from './document.controller';
 import { DocumentService } from './document.service';
 import { CrmModule } from '../crm/crm.module';
+import { SharepointModule } from '../sharepoint/sharepoint.module';
 
 @Module({
   imports: [
     ConfigModule,
     CrmModule,
+    SharepointModule,
   ],
   controllers: [DocumentController],
   providers: [
