+++ server/src/document/document.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, HttpStatus, HttpException } from '@nestjs/common';
 import { CrmService } from '../crm/crm.service';
+import { SharepointService } from '../sharepoint/sharepoint.service';
 
 const PACKAGE_VISIBILITY = {
   GENERAL_PUBLIC: 717170003,
@@ -33,163 +34,21 @@ function hyphenateGUID(unhyphenatedGUID) {
   ].join('');
 }
 
-// NOTE: There are no guarantees that Filed Land Use documents will show up until we have secured Sprint 10 enhancements from the EAS team.
-// This is because there have been instances that the dcpName portion for Filed LU is stripped of hyphens and spaces.
-// After that, we can also consider updating the app to read document locations.
-function stripDcpName(dcpName) {
-  return dcpName.replace(/'+/g, '').replace(/^\~|\#|\%|\&|\*|\{|\}|\\|\:|\<|\>|\?|\/|\||\"/g, '');
-}
-
-function constructFolderIdentifier(dcpName, recordId) {
-  return `${dcpName}_${recordId.toUpperCase().replace(/-/g, '')}`;
-}
-
 @Injectable()
 export class DocumentService {
   constructor(
     private readonly crmService: CrmService,
+    private readonly sharepointService: SharepointService,
   ){}
-
-  // We retrieve documents from the Sharepoint folder (`folderIdentifier` in the
-  // code below) that holds both documents from past revisions and the current
-  // revision. CRM automatically carries over documents from past revisions into
-  // this folder, and we deliberately upload documents for the latest/current
-  // revision into this folder.
-  async findPackageSharepointDocuments(packageName, id: string) {
-    try {
-      const strippedPackageName = stripDcpName(packageName);
-      const folderIdentifier = constructFolderIdentifier(strippedPackageName, id);
-
-      const { value: documents } = await this.crmService.getSharepointFolderFiles(`dcp_package/${folderIdentifier}`);
-
-      if (documents) {
-        return documents.map(document => ({
-          name: document['Name'],
-          timeCreated: document['TimeCreated'],
-          serverRelativeUrl: document['ServerRelativeUrl'],
-        }));
-      }
-
-      return [];
-    } catch (e) {
-      // Relay errors from crmService 
-      if (e instanceof HttpException) {
-        throw e;
-      } else {
-        const errorMessage = `An error occured while constructing and looking up folder for package. Perhaps the package name or id is wrong. ${JSON.stringify(e)}`;
-        console.log(errorMessage);
-
-        throw new HttpException({
-          code: 'SHAREPOINT_FOLDER_ERROR',
-          title: 'Bad Sharepoint folder lookup',
-          detail: errorMessage,
-          meta: {
-            packageName: packageName,
-            packageId: id,
-          }
-        }, HttpStatus.INTERNAL_SERVER_ERROR);
-      }
-    }
+  // NOTE: There are no guarantees that Filed Land Use documents will show up until we have secured Sprint 10 enhancements from the EAS team.
+  // This is because there have been instances that the dcpName portion for Filed LU is stripped of hyphens and spaces.
+  // After that, we can also consider updating the app to read document locations.
+  public stripDcpName(dcpName) {
+    return dcpName.replace(/'+/g, '').replace(/^\~|\#|\%|\&|\*|\{|\}|\\|\:|\<|\>|\?|\/|\||\"/g, '');
   }
 
-  async findArtifactSharepointDocuments(artifactName, id: string) {
-    try {
-      const strippedArtifactName = stripDcpName(artifactName);
-      const folderIdentifier = constructFolderIdentifier(strippedArtifactName, id);
-
-      const { value: documents } = await this.crmService.getSharepointFolderFiles(`dcp_artifacts/${folderIdentifier}`);
-
-      if (documents) {
-        return documents.map(document => ({
-          name: document['Name'],
-          timeCreated: document['TimeCreated'],
-          serverRelativeUrl: document['ServerRelativeUrl'],
-        }));
-      }
-
-      return [];
-    } catch (e) {
-      // Relay errors from crmService 
-      if (e instanceof HttpException) {
-        throw e;
-      } else {
-        const errorMessage = `An error occured while constructing and looking up folder for artifact. Perhaps the artifact name or id is wrong. ${JSON.stringify(e)}`;
-        console.log(errorMessage);
-
-        throw new HttpException({
-          code: 'SHAREPOINT_FOLDER_ERROR',
-          title: 'Bad Sharepoint folder lookup',
-          detail: errorMessage,
-          meta: {
-            artifactName: artifactName,
-            artifactsId: id,
-          }
-        }, HttpStatus.INTERNAL_SERVER_ERROR);
-      }
-    }
-  }
-
-
-  /**
-   * Injects associated documents into the given package
-   *
-   * @param      {Object{ dcp_name, dcp_packageid }} CRM Package with
-   *              dcp_name and dcp_packageid properties
-   * @return     {[Object]} The CRM Package with the 'documents' property hydrated,
-   *              if associated documents exist in CRM.
-   */
-  public async packageWithDocuments(projectPackage: any) {
-    const {
-      dcp_name,
-      dcp_packageid,
-    } = projectPackage;
-
-    try {
-      return {
-        ...projectPackage,
-        documents: await this.findPackageSharepointDocuments(dcp_name, dcp_packageid),
-      };
-    } catch (e) {
-      const errorMessage = `Error loading documents for package ${dcp_name}. ${JSON.stringify(e)}`;
-      console.log(errorMessage);
-
-      throw new HttpException({
-        "code": "PACKAGE_WITH_DOCUMENTS",
-        "title": "Package Documents Error",
-        "detail": errorMessage,
-      }, HttpStatus.NOT_FOUND);
-    }
-  }
-
-  /**
-   * Injects associated documents into the given artifact
-   *
-   * @param      {Object{ dcp_name, dcp_artifactsid }} CRM Artifact with
-   *              dcp_name and dcp_artifactsid properties
-   * @return     {[Object]} The CRM Artifact with the 'documents' property hydrated,
-   *              if associated documents exist in CRM.
-   */
-  public async artifactWithDocuments(projectArtifact: any) {
-    const {
-      dcp_name,
-      dcp_artifactsid,
-    } = projectArtifact;
-
-    try {
-      return {
-        ...projectArtifact,
-        documents: await this.findArtifactSharepointDocuments(dcp_name, dcp_artifactsid),
-      };
-    } catch (e) {
-      const errorMessage = `Error loading documents for artifact ${dcp_name}. ${JSON.stringify(e)}`;
-      console.log(errorMessage);
-
-      throw new HttpException({
-        "code": "PACKAGE_WITH_DOCUMENTS",
-        "title": "Artifact Documents Error",
-        "detail": errorMessage,
-      }, HttpStatus.NOT_FOUND);
-    }
+  public constructFolderIdentifier(dcpName, recordId) {
+    return `${dcpName}_${recordId.toUpperCase().replace(/-/g, '')}`;
   }
 
   // "path" refers to the "relative server path", the path
@@ -263,7 +122,7 @@ export class DocumentService {
         }
       }
 
-      return await this.crmService.getSharepointFile(pathSegment);
+      return await this.sharepointService.getSharepointFile(pathSegment);
     } catch(e) {
       const errorMessage = `Unable to provide document access. ${JSON.stringify(e)}`;
       console.log(errorMessage);
